extract common utils in one file
