+++ admin-panel/index.html
@@ -7,6 +7,7 @@
     <link rel="icon" href="/images/index.ico" type="image/x-icon" />
     <title>Admin Panel | Real Dev Squad</title>
     <script src="/constants.js"></script>
+    <script src="/utils.js"></script>
     <script src="/admin-panel/index.js" type="module" defer></script>
     <link rel="stylesheet" href="/admin-panel/index.css" />
   </head>
