+++ profile/index.html
@@ -9,6 +9,7 @@
     <link rel="stylesheet" href="/global.css" />
     <link rel="stylesheet" href="/profile/style.css" />
     <script src="/constants.js"></script>
+    <script src="/utils.js"></script>
     <script type="module" src="/profile/script.js" defer></script>
   </head>
