+++ utils.js
@@ -46,3 +46,37 @@ function addLoader(container) {
 function removeLoader(classname) {
   document.querySelector(`.${classname}`).remove();
 }
+
+async function makeApiCall(
+  url,
+  method = 'get',
+  body = null,
+  headers = [],
+  options = null,
+) {
+  try {
+    const response = await fetch(url, {
+      method,
+      body,
+      headers,
+      ...options,
+    });
+    return response;
+  } catch (err) {
+    console.error(MESSAGE_SOMETHING_WENT_WRONG, err);
+    return err;
+  }
+}
+
+function debounce(func, delay) {
+  let timerId;
+  return (...args) => {
+    if (timerId) {
+      clearTimeout(timerId);
+    }
+
+    timerId = setTimeout(() => {
+      func(...args);
+    }, delay);
+  };
+}
